Updated models to match database

+++ backend/Springboot/src/main/java/org/swi_project/repositories/CharacterRepository.java
@@ -3,5 +3,5 @@ package org.swi_project.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.swi_project.models.Character;
 
-public interface CharacterRepository extends JpaRepository<Character, Long> {
+public interface CharacterRepository extends JpaRepository<Character, Integer> {
 }
